[DO NOT MERGE] Add setting for power menu in lock screen privacy
Adds a new Setting for the user to show or hide power menu content
(cards & passes, controls) when the device is locked (secured). A value
of 0 (default) means hide content and a value of 1 means show.

The setting updates from the notification in the following way:
* If Secure.LOCK_SCREEN_SHOW_NOTIFICATIONS is 1 AND
Secure.LOCK_SCREEN_ALLOW_PRIVATE_NOTIFICATIONS is 1, the new setting is
set to 1 (the user has allowed sensitive content).
* If at least one of those settings is 0, the new setting is set to 0
(the user has not allowed sensitive content).
* If the settings are not set, the new setting is not set.

Test: build
Test: manual with adb
Bug: 152212779
Change-Id: If6d9436f633bc45330956372ea7f8a40ed22ed52