Calculate oldPaths for out of date LoadedApks
This was originally omitted because the only known case this
branch was hit was after a package update, which would inherently
change the code paths for the package.

Thus, spending the time to calculate these oldPaths was worthless
because they would never be applicable.

Seemingly this case where the paths are not changed has been
found, causing duplicate paths to be added to the same ClassLoader.
This warrants further investigation to determine why, but for now,
this should be safe to commit to fix the issue.

Bug: 149410951

Test: manual used steps at b/149410951#comment57, verified that
    paths were not added multiple times; device functions as normal
    otherwise

Change-Id: I632994dd19cae2b0101d7386826c8f8c7251051a