Refine logic about updateHideNavInputEventReceiver
The control target of navigation bar can be a non-window object. Non-
window control target will never request non-default insets state. So if
nav bar control target is not a window, we use the default visibility as
the requested visibility. This CL also obtains the behavior from the
controlling window instead of the focused window.

Fix: 154716165
Fix: 155297912
Test: atest DisplayPolicyTests
Change-Id: I34d7e0f3790a671fc89f5ef4b84d5848b94b81e4