applySyncTransaction: Add to syncSet earlier
We need to add modified WindowContainers to the BLASTSync set
before applying any changes. Recall that while participating in
sync window containers stop using the Display global
pending-transaction and instead redirect their changes in to the
BLASTSync-transaction. If we add to the BLAST sync too late, some
changes could go in to the WM transaction while others go in to
the BLASTSync-transaction. This means the effects would not be
applied atomically.

Bug: 149946388
Bug: 155814541
Test: Expand PIP. No flash.
Change-Id: I9b9b8f2308098d045a44004e2cca6dcbf84efcd9